ingest/requirements.txt
dlt
pandas
pyarrow
python-dotenv
tzdata